+++ Data/Keystones.lua
@@ -11,4 +11,5 @@ Data.keystones = {
   {seasonID = 13, seasonDisplayID = 1, itemID = 180653},
   {seasonID = 14, seasonDisplayID = 2, itemID = 180653},
   {seasonID = 15, seasonDisplayID = 3, itemID = 180653},
+  {seasonID = 17, seasonDisplayID = 1, itemID = 180653},
 }